[Keras] Move BatchMatmul class to backend.internal
