+++ python/flexflow/keras/models/base_model.py
@@ -18,7 +18,7 @@ from flexflow.core.flexflow_logger import fflogger
 
 from .tensor import Tensor
 from flexflow.keras.layers import Conv2D, Pooling2D, Flatten, Dense, Activation, Concatenate, Add, Subtract, Multiply, Dropout, BatchNormalization, Embedding, Reshape
-from flexflow.keras.backend import BatchMatmul
+from flexflow.keras.backend.internal import BatchMatmul
 from flexflow.keras.optimizers import SGD, Adam
 from flexflow.keras.callbacks import Callback, LearningRateScheduler, VerifyMetrics, EpochVerifyMetrics
 from flexflow.keras import losses as keras_losses
